move gcp resource deletion methods into utils file, put role set method into util
